Further breakdown of monolith functions
